misc/reporting-diff/requirements.txt
git+https://github.com/mansueto-institute/adaptive-lockdown@master#egg=adaptive
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
google-cloud-storage
oauth2client
pandas